refactor: Remove event config modal in favor of configuration page
