+++ products/customer_analytics/frontend/components/CustomerAnalyticsQueryCard.tsx
@@ -6,6 +6,7 @@ import { CardMeta } from 'lib/components/Cards/CardMeta'
 import { InsightMetaContent } from 'lib/components/Cards/InsightCard/InsightMeta'
 import { QueryCard } from 'lib/components/Cards/InsightCard/QueryCard'
 import { TopHeading } from 'lib/components/Cards/InsightCard/TopHeading'
+import { urls } from 'scenes/urls'
 
 import { QuerySchema } from '~/queries/schema/schema-general'
 import { InsightLogicProps } from '~/types'
@@ -16,8 +17,8 @@ import {
 } from 'products/customer_analytics/frontend/customerAnalyticsSceneLogic'
 
 import { CUSTOMER_ANALYTICS_DATA_COLLECTION_NODE_ID, SERIES_TO_EVENT_NAME_MAPPING } from '../constants'
+import { customerAnalyticsDashboardEventsLogic } from '../scenes/CustomerAnalyticsConfigurationScene/events/customerAnalyticsDashboardEventsLogic'
 import { buildDashboardItemId } from '../utils'
-import { eventConfigModalLogic } from './Insights/eventConfigModalLogic'
 
 interface CustomerAnalyticsQueryCardProps {
     insight: InsightDefinition
@@ -41,7 +42,7 @@ function generateUniqueKey(name: string, tabId: string, businessType: string, gr
 
 export function CustomerAnalyticsQueryCard({ insight, tabId }: CustomerAnalyticsQueryCardProps): JSX.Element {
     const { businessType, selectedGroupType } = useValues(customerAnalyticsSceneLogic)
-    const { addEventToHighlight, toggleModalOpen } = useActions(eventConfigModalLogic)
+    const { addEventToHighlight } = useActions(customerAnalyticsDashboardEventsLogic)
     const needsConfig = insight?.requiredSeries ? anyValueIsNull(insight.requiredSeries) : false
     const uniqueKey = generateUniqueKey(insight.name, tabId || '', businessType, selectedGroupType)
     const insightProps: InsightLogicProps<QuerySchema> = {
@@ -57,7 +58,6 @@ export function CustomerAnalyticsQueryCard({ insight, tabId }: CustomerAnalytics
             missingSeries.forEach((seriesName) => {
                 addEventToHighlight(SERIES_TO_EVENT_NAME_MAPPING[seriesName])
             })
-            toggleModalOpen()
         }
 
         return (
@@ -70,7 +70,7 @@ export function CustomerAnalyticsQueryCard({ insight, tabId }: CustomerAnalytics
                 <LemonBanner type="warning">
                     This insight requires configuration.
                     <div className="flex flex-row items-center gap-4 mt-2 max-w-160">
-                        <LemonButton type="primary" onClick={handleClick}>
+                        <LemonButton to={urls.customerAnalyticsConfiguration()} type="primary" onClick={handleClick}>
                             Configure events
                         </LemonButton>
                     </div>
